Refactoring the tests for making them more robust [2/n]
- Updated all test files to reference the newly separated download_history module
- Fixed import statements to correctly import HISTORY_MAGIC and DownloadHistoryFile class
  from synology_office_exporter.download_history
- Updated patch paths in tests to reflect the new module structure
- Modified tests in test_file_lock.py to directly access the DownloadHistoryFile class
