+++ tests/test_file_lock.py
@@ -9,8 +9,9 @@ from unittest.mock import patch, MagicMock
 from filelock import Timeout
 from io import StringIO
 
-from synology_office_exporter.exporter import SynologyOfficeExporter
+from synology_office_exporter.download_history import DownloadHistoryFile
 from synology_office_exporter.exception import DownloadHistoryError
+from synology_office_exporter.exporter import SynologyOfficeExporter
 
 
 class TestFileLock(unittest.TestCase):
@@ -28,43 +29,43 @@ class TestFileLock(unittest.TestCase):
 
     def test_enter_exit_acquires_and_releases_lock(self):
         """Test that __enter__ acquires a lock and __exit__ releases it."""
-        with patch('synology_office_exporter.exporter.FileLock') as mock_filelock:
+        with patch('synology_office_exporter.download_history.FileLock') as mock_filelock:
             mock_lock = MagicMock()
             mock_filelock.return_value = mock_lock
 
-            with SynologyOfficeExporter(self.synd_mock, output_dir=self.temp_dir.name, stat_buf=self.stat_buf):
-                # Verify lock was acquired
-                mock_lock.acquire.assert_called_once_with(blocking=False)
+            history_file = DownloadHistoryFile()
+            history_file.lock_history()
+            history_file.unlock_history()
 
-            # Verify lock was released
+            mock_filelock.assert_called_once()
+            mock_lock.acquire.assert_called_once_with(blocking=False)
             mock_lock.release.assert_called_once()
 
     def test_timeout_exception_when_lock_exists(self):
         """Test that a DownloadHistoryError is raised when the lock cannot be acquired."""
-        with patch('synology_office_exporter.exporter.FileLock') as mock_filelock:
+        with patch('synology_office_exporter.download_history.FileLock') as mock_filelock:
             mock_lock = MagicMock()
-            mock_lock.acquire.side_effect = Timeout("Lock already held")
             mock_filelock.return_value = mock_lock
+            mock_lock.acquire.side_effect = Timeout("lock_file")
 
-            with self.assertRaises(DownloadHistoryError) as cm:
-                with SynologyOfficeExporter(self.synd_mock, output_dir=self.temp_dir.name, stat_buf=self.stat_buf):
-                    self.assertTrue(False)  # Should not reach here
-
-            self.assertIn("Another process may be running", str(cm.exception))
+            history_file = DownloadHistoryFile()
+            with self.assertRaises(DownloadHistoryError):
+                history_file.lock_history()
 
     def test_lock_release_on_exception(self):
         """Test that lock is released even when an exception occurs within the context."""
-        with patch('synology_office_exporter.exporter.FileLock') as mock_filelock:
+        with patch('synology_office_exporter.download_history.FileLock') as mock_filelock, \
+                patch('synology_office_exporter.download_history.DownloadHistoryFile.load_history'), \
+                patch('synology_office_exporter.download_history.DownloadHistoryFile.save_history'):
             mock_lock = MagicMock()
             mock_filelock.return_value = mock_lock
 
             try:
-                with SynologyOfficeExporter(self.synd_mock, output_dir=self.temp_dir.name, stat_buf=self.stat_buf):
-                    raise ValueError("Test exception")
-            except ValueError:
-                pass  # Expected exception
+                with DownloadHistoryFile():
+                    raise RuntimeError("Test exception")
+            except RuntimeError:
+                pass
 
-            # Verify lock was still released despite the exception
             mock_lock.release.assert_called_once()
 
 
@@ -80,7 +81,7 @@ class TestFileLockIntegration(unittest.TestCase):
         """Clean up temporary directory."""
         self.temp_dir.cleanup()
 
-    @patch('synology_office_exporter.exporter.DownloadHistoryFile.load_history')
+    @patch('synology_office_exporter.download_history.DownloadHistoryFile.load_history')
     def test_concurrent_access_prevention(self, mock_load):
         """Test that a second instance cannot acquire the lock when first one holds it."""
         # First instance
